Eliminate duplicate solutions by hashing
Some solutions are formed differently, but because of the AAAAA piece,
visually look the same (the AAAAA piece only has two effective
orientations, but does not get any special handling).

This uses hashing of the display output to compare boards. If they'll
display the same, they are the same. Storing in HashSet prevents
duplicate boards from being stored.